fixed quest system bug and added bootleg cutscene

+++ Assets/Scripts/FireBoss/BucketPickup.cs
@@ -226,7 +226,7 @@ public class BucketPickup : MonoBehaviour
         {
             damage = 10f;
         }
-        fireBossManager.currentHealth -= damage;
+        fireBossManager.currentHealth -= (int)damage;
     }
 
     void PickupBucket()
